Added API impl to source file
